+++ opentaxforms/db.py
@@ -1,5 +1,6 @@
 
 import ut
+import config
 from sqlalchemy import MetaData, create_engine, select
 from sqlalchemy.exc import ProgrammingError
 from itertools import chain
@@ -28,7 +29,9 @@ def connect(appname,**kw):
 def connect_(**kw):
     # consumes keys from kw: user pw db
     global conn,engine,metadata
-    from config import cfg
+    cfg,log=config.setup(**kw)
+    if 'dirName' in kw:
+        del kw['dirName']
     usepostgres=kw.get('postgres',False) if cfg is None else cfg.postgres
     if usepostgres:
         kw.setdefault('user','postgres')