+++ test/run_fasttests.sh
@@ -1,4 +1,6 @@
 
-# run fast tests
-pytest
+# run faster tests
+PYTHONPATH=.. python test_opentaxforms.py -q
+# or can simply run:
+# pytest