+++ test/test_opentaxforms.py
@@ -1,49 +1,96 @@
 #! /usr/bin/env python
 
+# 'Otf' below [eg TestOtfSteps]  is 'OpenTaxForms'
+
+import os
+from os.path import join as pathjoin
+from shutil import copy
 from opentaxforms import ut
 
 class TestOtfBase(object):
     def setup_method(self, _):
-        self.dbpath='sqlite:///'+ut.Resource('test','opentaxforms.sqlite3').path()
+        self.testdir=ut.Resource('test','').path()
+        dbpath='sqlite:///'+self.testdir+'/opentaxforms.sqlite3'
+        self.defaultArgs=dict(
+            # skip cleanupFiles to allow comparison with target output
+            #skip=['c'],
+            # todo change dbpath to dburl
+            dbpath=dbpath,
+            )
     def teardown_method(self, _):
         pass
-class TestOtfSteps(TestOtfBase):
-    '''
-        these tests actually run the script
-        they dont start with 'test_' because
-        they each take several seconds to run
-        '''
-    def run_1040(self):
+    def ensureDir(self,folder):
+        if not ut.exists(folder):
+            os.makedirs(folder)
+    def run(self,**kw):
+        rootForms=kw.get('rootForms') or ['1040']
+        filesToCheck=kw.get('filesToCheck') or ['f%s-p1.html'%(form,) for form in rootForms]
+        kw.update(self.defaultArgs)
+        self.dirName=kw['dirName']
+        inputpdf=pathjoin(self.testdir,'forms-common','f1040.pdf')
+        outdir=pathjoin(self.testdir,self.dirName,'')
+        self.ensureDir(outdir)
+        copy(inputpdf,outdir)
         from opentaxforms import opentaxforms as otf
-        # skip cleanupFiles to allow comparison with target output
         returnval=otf.opentaxforms(
-            dirName='forms',rootForms=['1040'],
-            okToDownload=False,skip=['c'],
-            # todo change dbpath to dburl
-            dbpath=self.dbpath,
-            )
+            okToDownload=False,
+            **kw)
         if returnval!=0:
             raise Exception('run failed, no output to compare against target')
         import filecmp
         shallow=False
-        fileToCheck='f1040-p1.html'
-        outdir,targetdir='forms/','forms-targetOutput/'
-        filesMatch=filecmp.cmp(outdir+fileToCheck,targetdir+fileToCheck,shallow)
+        targetdir=pathjoin(self.testdir,'forms-targetOutput','')
         def fmtmsg(result,verb):
             return '{}: output file "{}" in "{}" {} target in "{}"'.format(
                 result,fileToCheck,outdir,verb,targetdir)
-        if filesMatch:
-            result,verb='PASS','matches'
-            print fmtmsg(result,verb)
-        else:
-            result,verb='FAIL','does NOT match'
-            raise Exception(fmtmsg(result,verb))
+        for fileToCheck in filesToCheck:
+            filesMatch=filecmp.cmp(
+                pathjoin(outdir,fileToCheck),
+                pathjoin(targetdir,fileToCheck),
+                shallow)
+            if filesMatch:
+                result,verb='PASS','matches'
+                print fmtmsg(result,verb)
+            else:
+                result,verb='FAIL','does NOT match'
+                raise Exception(fmtmsg(result,verb))
+class TestOtfSteps(TestOtfBase):
+    '''
+        These 'steps' tests actually run the script.
+        run_1040_full runs 'all steps' and thus doesnt
+        start with 'test_' because it runs for several seconds
+        '''
+    # todo use a less complex form than 1040 to speed testing
+    def run_1040_full(self):
+        self.run(
+            rootForms=['1040'],
+            filesToCheck=['f1040-p1.html'],
+            dirName='forms_1040_full',
+            ignoreCaches=True,
+            )
+    def test_run_1040_xfa(self):
+        dirName=pathjoin(self.testdir,'forms_1040_xfa')
+        self.ensureDir(dirName)
+        # use cached pdf info to speed the run
+        pdfinfo=pathjoin(self.testdir,'forms-common','f1040-pdfinfo.pickl')
+        copy(pdfinfo,dirName)
+        self.run(
+            rootForms=['1040'],
+            filesToCheck=['f1040-fmt.xml'],
+            dirName=dirName,
+            steps=['x'],
+            # speeds testing
+            computeOverlap=False,
+            )
+    # todo add tests of further steps,
+    # todo   made fast via pickled results of previous step
 
 class TestOtfApiBase(object):
     def setup_method(self, _):
         from opentaxforms.serve import createApp
         dbpath='sqlite:///'+ut.Resource('test','opentaxforms.sqlite3').path()
-        self.app=createApp(dbpath=dbpath)
+        # dirName=None means dont look for a forms/ directory
+        self.app=createApp(dbpath=dbpath,dirName=None)
         self.client=self.app.test_client()
     def teardown_method(self, _):
         pass
@@ -79,9 +126,20 @@ class TestOtfApi(TestOtfApiBase):
 
 def main(args):
     def usage():
-        print 'usage: "%s -f" for fasttests or "%s -s" for slow tests'%(args[0],args[0])
+        print 'usage: "%s [-q|-s|-f|-x|-a]"\n-q=quick script tests\n-s=slow script tests\n-f=full 1040\n-x=xfa-only 1040\n-a=api tests'%(args[0],)
     if len(args)==2:
-        if args[1]=='-f':
+        if args[1] in ('-q','-s','-f','-x'):
+            testRunner=TestOtfSteps()
+            testRunner.setup_method(0)
+            if args[1]=='-q':
+                testRunner.test_run_1040_xfa()
+            elif args[1]=='-s':
+                testRunner.run_1040_full()
+            elif args[1]=='-x':
+                testRunner.test_run_1040_xfa()
+            elif args[1]=='-f':
+                testRunner.run_1040_full()
+        elif args[1]=='-a': # run api tests
             testRunner=TestOtfApi()
             testRunner.setup_method(0)
 
@@ -90,11 +148,6 @@ def main(args):
             testRunner.test_api_noresults()
             testRunner.test_api_filterslots()
 
-        elif args[1]=='-s': # run slow tests
-            testRunner=TestOtfSteps()
-            testRunner.setup_method(0)
-
-            testRunner.run_1040()
         else:
             usage()