+++ CHANGES.md
@@ -1,3 +1,21 @@
+0.4.15 (2017-01-09)
+
+- bugfix: the ignoreCaches option now works.
+- bugfix: skip forms setup if dirName is None.
+- bugfix: ut.exists now insensitive to trailing slash.
+- test: moved from relative to absolute paths so eg py.test works from outside of test/ dir.
+- test: reorganize test/ dir: test_* subdirs are disposable, test-* arent.
+- test/test_opentaxforms.py: command line options changed and expanded.
+- test/test_opentaxforms.py: refactor TestOtfSteps, add test_run_1040_xfa, and finally add a script test to the [pytest] automated tests.
+- test/forms/f1040.js,test/run_html5.sh: add casperjs script [to test f1040 html files] and shell script [for context].
+- add test/README.md as guide to the test/ directory.
+- probe access to phantomjs,casperjs in .travis.yml, hopefully for future integration of html5 tests.
+- opentaxforms.py: in main, move form-summarizing code to try-else block.
+- cache results of Form.pdfInfo for faster test run.
+- add internal option to computeOverlap [for layout boxes; currently done inefficiently, thus costly].
+- update 1040 target output file to fix broken test.
+- cleanup opentaxforms.py/main.
+
 0.4.14 (2016-12-31)
 
 - mv math.py cmds.py: workaround for tox build failure [tox.Random imports py.math and gets our math.py instead].